split CCConfig and Capabilities config changed path to capabilities config object in JS renderer
